+++ README.md
@@ -147,5 +147,5 @@
   - Order로 수정했습니다. 
   - 구현하지는 못했지만 (A, B, C 물품을 구매한 주문 ㄱ), (B, C, D를 구매한 주문 ㄴ)이 있다면 ㄱ과 ㄴ은 Order이 되어야 하기 때문에 Order가 맞다고 이해했습니다.
 - [x] 도메인과 service에 대한 테스트가 없는 것들이 있는것 같아요!
-- [ ] Impl을 붙이는 방식은 이제 사용을 권장하지 않는 것 같아서요. 
+- [x] Impl을 붙이는 방식은 이제 사용을 권장하지 않는 것 같아서요. 
   - CartItemRepository를 interface로 하고 CartItemDao가 implments하는 방식으로 해도되지 않을까 싶습니다
