+++ src/main/java/woowacourse/shoppingcart/application/CartService.java
@@ -5,9 +5,9 @@ import java.util.List;
 import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import woowacourse.shoppingcart.dao.CartItemDao;
-import woowacourse.shoppingcart.dao.CustomerDao;
-import woowacourse.shoppingcart.dao.ProductDao;
+import woowacourse.shoppingcart.dao.CartItemRepository;
+import woowacourse.shoppingcart.dao.CustomerRepository;
+import woowacourse.shoppingcart.dao.ProductRepository;
 import woowacourse.shoppingcart.domain.Cart;
 import woowacourse.shoppingcart.domain.Product;
 import woowacourse.shoppingcart.dto.AddCartItemRequest;
@@ -25,11 +25,11 @@ import woowacourse.shoppingcart.exception.NotInCustomerCartItemException;
 @Transactional(rollbackFor = Exception.class)
 public class CartService {
 
-    private final CartItemDao cartItemDao;
-    private final CustomerDao customerDao;
-    private final ProductDao productDao;
+    private final CartItemRepository cartItemDao;
+    private final CustomerRepository customerDao;
+    private final ProductRepository productDao;
 
-    public CartService(final CartItemDao cartItemDao, final CustomerDao customerDao, final ProductDao productDao) {
+    public CartService(final CartItemRepository cartItemDao, final CustomerRepository customerDao, final ProductRepository productDao) {
         this.cartItemDao = cartItemDao;
         this.customerDao = customerDao;
         this.productDao = productDao;
