+++ src/test/java/woowacourse/shoppingcart/dao/CartItemDaoTest.java
@@ -21,15 +21,15 @@ import woowacourse.shoppingcart.domain.Product;
 @AutoConfigureTestDatabase(replace = Replace.NONE)
 @Sql(scripts = {"classpath:schema.sql", "classpath:data.sql"})
 @TestConstructor(autowireMode = TestConstructor.AutowireMode.ALL)
-public class CartItemDaoImplTest {
-    private final CartItemDao cartItemDao;
-    private final ProductDao productDao;
+public class CartItemDaoTest {
+    private final CartItemRepository cartItemDao;
+    private final ProductRepository productDao;
     private final JdbcTemplate jdbcTemplate;
 
-    public CartItemDaoImplTest(JdbcTemplate jdbcTemplate) {
+    public CartItemDaoTest(JdbcTemplate jdbcTemplate) {
         this.jdbcTemplate = jdbcTemplate;
-        cartItemDao = new CartItemDaoImpl(jdbcTemplate);
-        productDao = new ProductDaoImpl(jdbcTemplate);
+        cartItemDao = new CartItemDao(jdbcTemplate);
+        productDao = new ProductDao(jdbcTemplate);
     }
 
     @BeforeEach
